backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
aiolimiter==1.1.0
cryptography==43.0.1
anthropic==0.40.0
reportlab==4.2.2
pillow==10.4.0
pytesseract==0.3.13
python-multipart==0.0.12
alembic==1.13.3
psycopg2-binary==2.9.9
google-generativeai==0.8.3
